requirements.txt
pandas==1.4.2
streamlit==1.8.1
Pillow==9.0.1

